trying edge collapse again, starting from slow but robust algorithms
